fix(mcp): report the brain release next to the MCP server's own version
The unauthenticated probe answered `version: 0.3.0` while /health on the
same host answered `2.2.0`. Both are correct — one is the MCP server, the
other the product — but two different answers to "which brain am I
talking to" cost debugging time during setup, which is exactly when the
probe is used.

The probe now carries `serviceVersion` alongside `version`, reading the
same package.json constant /health does, extracted to common/ so the two
surfaces cannot drift.
